made changes for rng_model to properly import pecos-rng crate
